requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
backcall==0.2.0
billiard==4.1.0
blinker==1.6.3
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.6
dvc==3.26.0
dvc-data==2.18.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
ensure==1.0.3
entrypoints==0.4
executing==2.0.0
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
funcy==2.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grandalf==0.8
greenlet==3.0.0
grpcio==1.59.0
gto==1.4.0
h5py==3.10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.14.0
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.4.5
kombu==5.3.2
libclang==16.0.6
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mlflow==2.7.1
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.9
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pygit2==1.13.1
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
python-box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.1
scipy==1.11.3
scmrepo==1.3.1
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
sqltrie==0.8.0
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
voluptuous==0.13.1
waitress==2.1.2
wcwidth==0.2.8
websocket-client==1.6.4
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
-e .